move component and add suspence
